Bug of the year! While the old logic _seems_ very sane and intuitive, it is in fact wrong. getLengthSq relies on a quadratic multiplication (EG: x^2), which will treat negative numbers the same way as positive.
EG: length(23, 23) == length(-23, -23).

Now, if your binary tree completely relies on this "length" thing, the search algorithm may dive into the wrong branch of the tree, ergo it ends up searching on the far left side, while the real value could be on the right side. 

It would appear Microsoft their stdlib uses a (variant-of) depth-first (sitation needed), so it shall search the entire tree if no value is found (remember, depth-first can go back to the parent node), linux however, seems to halt searching once a leave node is found, it does not step back into the parent node. I'm 93% sure this is the problem.

As a fix I've changed the appropriate operators to produce a unique result based on the parameters given (see diff). This may feel less-intuitive, but prevents future bugs with other data structures that rely on the less than and greater than operators.